+++ plugins/gradus/src/controller/controller.cpp
@@ -244,6 +244,7 @@ void Controller::willClose([[maybe_unused]] VSTGUI::VST3Editor* editor)
     ringDataBridge_.clearLanes();
     ringDisplay_ = nullptr;
     detailStrip_ = nullptr;
+    laneSwingKnobs_.fill(nullptr);
     presetBrowserView_ = nullptr;
     savePresetDialogView_ = nullptr;
     activeEditor_ = nullptr;